MyApps-13: Move get book UseCase to submodules.
