+++ domain/di/src/main/java/xyz/torquato/myapps/domain/di/UseCaseImpl.kt
@@ -7,10 +7,12 @@ import dagger.hilt.components.SingletonComponent
 import xyz.torquato.myaps.domain.api.pagging.IPagingRepository
 import xyz.torquato.myaps.domain.api.web.ILocalWebRepository
 import xyz.torquato.myaps.domain.api.web.IQueryRepository
+import xyz.torquato.myaps.domain.api.web.IWebRepository
 import xyz.torquato.myaps.domain.api.web.cache.IWebCacheRepository
 import xyz.torquato.myaps.domain.impl.web.GetCacheBooksUseCase
 import xyz.torquato.myaps.domain.impl.web.GetSelectedBookUseCase
 import xyz.torquato.myaps.domain.impl.web.SetQueryUseCase
+import xyz.torquato.myaps.domain.impl.web.content.GetBooksUseCase
 import xyz.torquato.myaps.domain.impl.web.content.SetSelectedBookUseCase
 import xyz.torquato.myaps.domain.impl.web.paging.GetNextPageUseCase
 import javax.inject.Singleton
@@ -58,6 +60,20 @@ object UseCaseImpl {
     fun provideGetCacheBooksUseCase(
         webCacheRepository: IWebCacheRepository
     ): GetCacheBooksUseCase = GetCacheBooksUseCase(
+        webCacheRepository
+    )
+
+    @Singleton
+    @Provides
+    fun provideGetBooksUseCase(
+        queryRepository: IQueryRepository,
+        webRepository: IWebRepository,
+        pagingRepository: IPagingRepository,
+        webCacheRepository: IWebCacheRepository
+    ): GetBooksUseCase = GetBooksUseCase(
+        queryRepository,
+        webRepository,
+        pagingRepository,
         webCacheRepository
     )
 }
